refactor(project): reorganize internal packages without behavior changes
