+++ internal/api/handlers/response.go
@@ -1,0 +1,16 @@
+package handlers
+
+import (
+	"encoding/json"
+	"net/http"
+)
+
+func respond(w http.ResponseWriter, code int, data interface{}) {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(code)
+	json.NewEncoder(w).Encode(data) //nolint:errcheck
+}
+
+func respondError(w http.ResponseWriter, code int, msg string) {
+	respond(w, code, map[string]string{"error": msg})
+}
