+++ internal/api/router.go
@@ -3,25 +3,11 @@ package api
 import (
 	"net/http"
 	"strings"
+
+	"github.com/kirilltahmazidi/aggregator/internal/api/handlers"
 )
 
-// Middleware
-func enableCORS(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Access-Control-Allow-Origin", "*") // !
-		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, OPTIONS")
-		w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
-		// если это предзапрос - CORS (OPTIONS) => 200 OK
-		if r.Method == http.MethodOptions {
-			w.WriteHeader(http.StatusOK)
-			return
-		}
-
-		next.ServeHTTP(w, r)
-	})
-}
-
-func NewRouter(h *Handler) http.Handler {
+func NewRouter(h *handlers.Handler) http.Handler {
 	mux := http.NewServeMux()
 
 	mux.HandleFunc("/health", h.Health)
