+++ internal/store/store.go
@@ -7,23 +7,21 @@ import (
 	"time"
 
 	_ "github.com/jackc/pgx/v5/stdlib"
-	"github.com/lib/pq"
 )
 
-// OrderStatus — статус заказа на каждом этапе жизненного цикла
+// OrderStatus is the order lifecycle status.
 type OrderStatus string
 
 const (
-	StatusPending          OrderStatus = "pending"                        // ждёт исполнителя
-	StatusSearching        OrderStatus = "searching"                      // идёт поиск эксплуатанта
-	StatusMatched          OrderStatus = "matched"                        // исполнитель найден
-	StatusConfirmed        OrderStatus = "confirmed"                      // контракт подписан
-	StatusCompletedPending OrderStatus = "completed_pending_confirmation" // оператор сообщил об успехе, ждём подтверждения заказчика
-	StatusCompleted        OrderStatus = "completed"                      // заказ выполнен
-	StatusDispute          OrderStatus = "dispute"                        // открыт спор
+	StatusPending          OrderStatus = "pending"
+	StatusSearching        OrderStatus = "searching"
+	StatusMatched          OrderStatus = "matched"
+	StatusConfirmed        OrderStatus = "confirmed"
+	StatusCompletedPending OrderStatus = "completed_pending_confirmation"
+	StatusCompleted        OrderStatus = "completed"
+	StatusDispute          OrderStatus = "dispute"
 )
 
-// Order — запись о заказе
 type Order struct {
 	ID               string      `json:"id"`
 	CustomerID       string      `json:"customer_id"`
@@ -34,8 +32,8 @@ type Order struct {
 	ToLat            float64     `json:"to_lat"`
 	ToLon            float64     `json:"to_lon"`
 	Status           OrderStatus `json:"status"`
-	OperatorID       string      `json:"operator_id,omitempty"`   // заполняется когда эксплуатант даёт оферту
-	OfferedPrice     float64     `json:"offered_price,omitempty"` // цена от эксплуатанта
+	OperatorID       string      `json:"operator_id,omitempty"`
+	OfferedPrice     float64     `json:"offered_price,omitempty"`
 	MissionType      string      `json:"mission_type"`
 	SecurityGoals    []string    `json:"security_goals,omitempty"`
 	TopLeftLat       float64     `json:"top_left_lat,omitempty"`
@@ -47,7 +45,6 @@ type Order struct {
 	CreatedAt        time.Time   `json:"created_at"`
 }
 
-// Operator — зарегистрированный эксплуатант
 type Operator struct {
 	ID           string `json:"id"`
 	Name         string `json:"name"`
@@ -56,7 +53,6 @@ type Operator struct {
 	PasswordHash string `json:"-"`
 }
 
-// Customer — зарегистрированный заказчик
 type Customer struct {
 	ID           string `json:"id"`
 	Name         string `json:"name"`
@@ -65,20 +61,17 @@ type Customer struct {
 	PasswordHash string `json:"-"`
 }
 
-// Store — хранилище на основе PostgreSQL
+// Store wraps PostgreSQL access.
 type Store struct {
-	db *sql.DB // пул соединений к базе данных
+	db *sql.DB
 }
 
-// New открывает соединение с базой данных и проверяет что оно живое.
 func New(databaseURL string) (*Store, error) {
-	// просто создаёт пул
 	db, err := sql.Open("pgx", databaseURL)
 	if err != nil {
 		return nil, fmt.Errorf("sql.Open: %w", err)
 	}
 
-	// Ping — вот здесь происходит первое подключение
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
 	if err := db.PingContext(ctx); err != nil {
@@ -88,317 +81,11 @@ func New(databaseURL string) (*Store, error) {
 	return &Store{db: db}, nil
 }
 
-// Close закрывает пул соединений при остановке сервиса
 func (s *Store) Close() error {
 	return s.db.Close()
 }
 
-// RunMigrations выполняет SQL-схему — запускаем при старте сервиса.
 func (s *Store) RunMigrations(sqlText string) error {
 	_, err := s.db.Exec(sqlText)
 	return err
-}
-
-// Orders
-func (s *Store) SaveOrder(o *Order) error {
-	securityGoals := o.SecurityGoals
-	if securityGoals == nil {
-		securityGoals = []string{}
-	}
-
-	// ON CONFLICT (id) DO UPDATE — upsert: вставить или обновить если id уже есть
-	_, err := s.db.Exec(`
-		INSERT INTO orders (
-			id, customer_id, description, budget,
-			from_lat, from_lon, to_lat, to_lon,
-			status, operator_id, offered_price,
-			mission_type, security_goals,
-			top_left_lat, top_left_lon, bottom_right_lat, bottom_right_lon,
-			commission_amount, operator_amount,
-			created_at
-		)
-		VALUES ($1, $2, $3, $4,
-			$5, $6, $7, $8,
-			$9, $10, $11,
-			$12, $13,
-			$14, $15, $16, $17,
-			$18, $19,
-			$20)
-		ON CONFLICT (id) DO UPDATE SET status = EXCLUDED.status
-	`, o.ID, o.CustomerID, o.Description, o.Budget,
-		o.FromLat, o.FromLon, o.ToLat, o.ToLon,
-		string(o.Status), o.OperatorID, o.OfferedPrice,
-		o.MissionType, pq.Array(securityGoals),
-		o.TopLeftLat, o.TopLeftLon, o.BottomRightLat, o.BottomRightLon,
-		o.CommissionAmount, o.OperatorAmount,
-		o.CreatedAt)
-	return err
-}
-
-func (s *Store) GetOrder(id string) (*Order, bool) {
-	o := &Order{}
-	err := s.db.QueryRow(`
-		SELECT id, customer_id, description, budget,
-			from_lat, from_lon, to_lat, to_lon,
-			status, operator_id, offered_price,
-			mission_type, security_goals,
-			top_left_lat, top_left_lon, bottom_right_lat, bottom_right_lon,
-			commission_amount, operator_amount,
-			created_at
-		FROM orders WHERE id = $1
-	`, id).Scan(
-		&o.ID, &o.CustomerID, &o.Description, &o.Budget,
-		&o.FromLat, &o.FromLon, &o.ToLat, &o.ToLon,
-		&o.Status, &o.OperatorID, &o.OfferedPrice,
-		&o.MissionType, pq.Array(&o.SecurityGoals),
-		&o.TopLeftLat, &o.TopLeftLon, &o.BottomRightLat, &o.BottomRightLon,
-		&o.CommissionAmount, &o.OperatorAmount,
-		&o.CreatedAt,
-	)
-	if err == sql.ErrNoRows {
-		return nil, false // не нашли — возвращаем false как раньше
-	}
-	if err != nil {
-		return nil, false
-	}
-	return o, true
-}
-
-func (s *Store) ListOrders() []*Order {
-	rows, err := s.db.Query(`
-		SELECT id, customer_id, description, budget,
-			from_lat, from_lon, to_lat, to_lon,
-			status, operator_id, offered_price,
-			mission_type, security_goals,
-			top_left_lat, top_left_lon, bottom_right_lat, bottom_right_lon,
-			commission_amount, operator_amount,
-			created_at
-		FROM orders ORDER BY created_at DESC
-	`)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	var orders []*Order
-	for rows.Next() { // идём по строкам результата одна за одной
-		o := &Order{}
-		if err := rows.Scan(
-			&o.ID, &o.CustomerID, &o.Description, &o.Budget,
-			&o.FromLat, &o.FromLon, &o.ToLat, &o.ToLon,
-			&o.Status, &o.OperatorID, &o.OfferedPrice,
-			&o.MissionType, pq.Array(&o.SecurityGoals),
-			&o.TopLeftLat, &o.TopLeftLon, &o.BottomRightLat, &o.BottomRightLon,
-			&o.CommissionAmount, &o.OperatorAmount,
-			&o.CreatedAt,
-		); err != nil {
-			continue
-		}
-		orders = append(orders, o)
-	}
-	return orders
-}
-
-func (s *Store) ListOrdersByCustomer(customerID string) []*Order {
-	rows, err := s.db.Query(`
-		SELECT id, customer_id, description, budget,
-			from_lat, from_lon, to_lat, to_lon,
-			status, operator_id, offered_price,
-			mission_type, security_goals,
-			top_left_lat, top_left_lon, bottom_right_lat, bottom_right_lon,
-			commission_amount, operator_amount,
-			created_at
-		FROM orders WHERE customer_id = $1 ORDER BY created_at DESC
-	`, customerID)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	var orders []*Order
-	for rows.Next() {
-		o := &Order{}
-		if err := rows.Scan(
-			&o.ID, &o.CustomerID, &o.Description, &o.Budget,
-			&o.FromLat, &o.FromLon, &o.ToLat, &o.ToLon,
-			&o.Status, &o.OperatorID, &o.OfferedPrice,
-			&o.MissionType, pq.Array(&o.SecurityGoals),
-			&o.TopLeftLat, &o.TopLeftLon, &o.BottomRightLat, &o.BottomRightLon,
-			&o.CommissionAmount, &o.OperatorAmount,
-			&o.CreatedAt,
-		); err != nil {
-			continue
-		}
-		orders = append(orders, o)
-	}
-	return orders
-}
-
-func (s *Store) UpdateOrderStatus(id string, status OrderStatus) bool {
-	res, err := s.db.Exec(`UPDATE orders SET status = $1 WHERE id = $2`, string(status), id)
-	if err != nil {
-		return false
-	}
-	n, _ := res.RowsAffected() // сколько строк было изменено
-	return n > 0               // если 0 — такого заказа нет
-}
-
-// ConfirmPrice фиксирует выбор оператора, финальную цену и комиссию.
-func (s *Store) ConfirmPrice(id, operatorID string, acceptedPrice, commissionAmount float64) bool {
-	if acceptedPrice <= 0 {
-		return false
-	}
-	operatorAmount := acceptedPrice - commissionAmount
-	res, err := s.db.Exec(`
-		UPDATE orders 
-		SET status = $1, offered_price = $2, commission_amount = $3, operator_amount = $4
-		WHERE id = $5 AND status = $6 AND operator_id = $7
-	`, string(StatusConfirmed), acceptedPrice, commissionAmount, operatorAmount, id, string(StatusMatched), operatorID)
-
-	if err != nil {
-		return false
-	}
-	n, _ := res.RowsAffected()
-	return n > 0
-}
-
-// ConfirmCompletion фиксирует подтверждение выполнения заказчиком.
-func (s *Store) ConfirmCompletion(id string) bool {
-	res, err := s.db.Exec(`
-		UPDATE orders 
-		SET status = $1 
-		WHERE id = $2 AND status = $3
-	`, string(StatusCompleted), id, string(StatusCompletedPending))
-
-	if err != nil {
-		return false
-	}
-	n, _ := res.RowsAffected()
-	return n > 0
-}
-
-// SetOperatorOffer сохраняет оферту от эксплуатанта: его ID и предложенную цену.
-func (s *Store) SetOperatorOffer(orderID, operatorID string, price float64) bool {
-	if price <= 0 {
-		return false
-	}
-	res, err := s.db.Exec(`
-		UPDATE orders 
-		SET status = $1, operator_id = $2, offered_price = $3 
-		WHERE id = $4 AND status IN ($5, $6)
-	`, string(StatusMatched), operatorID, price, orderID, string(StatusSearching), string(StatusPending))
-
-	if err != nil {
-		return false
-	}
-	n, _ := res.RowsAffected()
-	return n > 0
-}
-
-// ProcessOrderResult обрабатывает результат выполнения заказа от оператора
-func (s *Store) ProcessOrderResult(orderID string, success bool) bool {
-	targetStatus := StatusDispute
-	if success {
-		targetStatus = StatusCompletedPending
-	}
-	res, err := s.db.Exec(`
-		UPDATE orders 
-		SET status = $1 
-		WHERE id = $2 AND status = $3
-	`, string(targetStatus), orderID, string(StatusConfirmed))
-
-	if err != nil {
-		return false
-	}
-	n, _ := res.RowsAffected()
-	return n > 0
-}
-
-// Operators
-
-func (s *Store) SaveOperator(op *Operator) error {
-	_, err := s.db.Exec(`
-		INSERT INTO operators (id, name, license, email, password_hash)
-		VALUES ($1, $2, $3, $4, $5)
-		ON CONFLICT (id) DO NOTHING
-	`, op.ID, op.Name, op.License, op.Email, op.PasswordHash)
-	return err
-}
-
-func (s *Store) GetOperator(id string) (*Operator, bool) {
-	op := &Operator{}
-	err := s.db.QueryRow(`
-		SELECT id, name, license, email, password_hash FROM operators WHERE id = $1
-	`, id).Scan(&op.ID, &op.Name, &op.License, &op.Email, &op.PasswordHash)
-	if err != nil {
-		return nil, false
-	}
-	return op, true
-}
-
-func (s *Store) GetOperatorByEmail(email string) (*Operator, bool) {
-	op := &Operator{}
-	err := s.db.QueryRow(`
-		SELECT id, name, license, email, password_hash FROM operators
-		WHERE LOWER(email) = LOWER($1)
-		ORDER BY created_at ASC
-		LIMIT 1
-	`, email).Scan(&op.ID, &op.Name, &op.License, &op.Email, &op.PasswordHash)
-	if err != nil {
-		return nil, false
-	}
-	return op, true
-}
-
-func (s *Store) SetOperatorPasswordHash(id, passwordHash string) bool {
-	res, err := s.db.Exec(`UPDATE operators SET password_hash = $1 WHERE id = $2`, passwordHash, id)
-	if err != nil {
-		return false
-	}
-	n, _ := res.RowsAffected()
-	return n > 0
-}
-
-// Customers
-func (s *Store) SaveCustomer(c *Customer) error {
-	_, err := s.db.Exec(`
-		INSERT INTO customers (id, name, email, phone, password_hash)
-		VALUES ($1, $2, $3, $4, $5)
-		ON CONFLICT (id) DO NOTHING
-	`, c.ID, c.Name, c.Email, c.Phone, c.PasswordHash)
-	return err
-}
-
-func (s *Store) GetCustomer(id string) (*Customer, bool) {
-	c := &Customer{}
-	err := s.db.QueryRow(`
-		SELECT id, name, email, phone, password_hash FROM customers WHERE id = $1
-	`, id).Scan(&c.ID, &c.Name, &c.Email, &c.Phone, &c.PasswordHash)
-	if err != nil {
-		return nil, false
-	}
-	return c, true
-}
-
-func (s *Store) GetCustomerByEmail(email string) (*Customer, bool) {
-	c := &Customer{}
-	err := s.db.QueryRow(`
-		SELECT id, name, email, phone, password_hash FROM customers
-		WHERE LOWER(email) = LOWER($1)
-		ORDER BY created_at ASC
-		LIMIT 1
-	`, email).Scan(&c.ID, &c.Name, &c.Email, &c.Phone, &c.PasswordHash)
-	if err != nil {
-		return nil, false
-	}
-	return c, true
-}
-
-func (s *Store) SetCustomerPasswordHash(id, passwordHash string) bool {
-	res, err := s.db.Exec(`UPDATE customers SET password_hash = $1 WHERE id = $2`, passwordHash, id)
-	if err != nil {
-		return false
-	}
-	n, _ := res.RowsAffected()
-	return n > 0
 }
